Build a Resource attribute map.
The attribute map is a taken from the JSON schema response portion,
given the definition found elsewhere in the JSON schema.

You may ask yourself, why can't I just use the definition portion
as-is? Great question! The definition does not have to map 1:1
to the actual expected resposne of the RESTful action. There
may be unused definitions, or mis-matching between definition names
and response attributes. So we must map in this way.